% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_samples.R
\name{get_norm_samples}
\alias{get_norm_samples}
\title{Get random normal samples for each element of list}
\usage{
get_norm_samples(mylist)
}
\arguments{
\item{mylist}{list of numeric vectors}
}
\value{
A list of random normal samples with same length as the list elements
}
\description{
Get random normal samples for each element of list
}
\examples{
mylist <- list(a = 1, b = 1:10, c = 1:20)
get_norm_samples(mylist)
}
